requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
pyjwt==2.9.0
structlog==24.4.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
torch
torchvision
pynvml==11.5.0
pydantic==2.9.2
